Incorporated pitch shift code into the ContourGenerator. The pitch shift code allows syllables to be annotated with a 'local_f0_shift' feature. This should make implementation of JSML 'prosody pitch' markup possible. Currently there does not seem to be any processing that sets local_f0_shift as of yet.
Voice now has a 'get/setPitchShift method that allows the setting of a
pitch shift. (1.0 means no shift), 2.0 means twice the pitch.

Updated the release notes to reflect the changes.

The .txt files have not really changed. I touched them to force the
binary forms of these files to build. I was hesitant to stop this cvs
commit since my experience is that trying to stop a commit half-way
through can lead to a stuck lock in the DB which can take SourceForge
several days to fix.

After applying these changes I ran the tests and all pass.