Flite 1.1 Upgrade.
Flite 1.1 uses a mini-lexicon for the limited-domain time voice.
FreeTTS has been updated to also include a miniature lexicon for this
voice.  This required a bit of restructuring. A lexicon, in reality
consists of three data sources, the compiled form, the addenda and the
letter to sound rules. There was some inconsistancy in how these three
files were named which made it difficult to easily switch between two
lexicons. I made all the sources consistent. They are of the form
basename + suffix where basename is either cmulex (the original
lexicon) or cmutime (the limited domain lexicon). The suffixes are
'_lts.txt' for the letter to sound rules, '_compiled.txt' for the
compiled form and '_addenda.txt' for the addenda.

The generatebinary programs were also modified to allow the generation
of binary forms of different lexicons (they were originally hardwired
to work on the cmulex lexicon).

Finally, a new Voice CMUTimeAWBVoice was created that specifically
uses the new mini-lexicon.

Using the new lexicon