Fixed the 'OutOfMemoryError' that was encountered when loading from the jar.
Problem Description:

There are two methods of loading the binary cluster unit database, via
mapped files and the traditional old binary I/O method. The mapped
method is the preferred method as it is almost always faster, however
it can only be used on real files. When the database is loaded from
jar file, the nio file mapping cannot be used so we fall back on the
old i/o method. There was a bug in loading of the database with the
old i/o in that it would read one field from the DB twice, causing
subsequent reads to be out of sync. One of the subsequent fields is
used to allocate an array of data.  Since the reads were out of sync,
instead of getting the actuall size (a few hundred elements), we were
trying to allocate several hundred million elements, resulting in the
out of memory error.  The error only occurs when loading the system
from jar files because that is the only time that we'll use the old io
style.

The bug was more difficult to find than usual because:

1) There seemed to be more than enough memory available when the
system failed.

2) The OutOfMemoryError does not fill in any stack trace information
(presumably because there may not be enough memory to hold that info),
leaving little trace of where the exception occurred.


Solution:
Once the erroneous duplicate read was removed, the MixedVoices demo
runs fine from the jar files.