Workaround for javax.sound linux crash.
The bug it is described here:

http://developer.java.sun.com/developer/bugParade/bugs/4498848.html

The bug seems to be related to synchronization of the close
call on the clip.  If we delay before the close by
a small amount, the crash is averted.  This is a WORKAROUND
only and should be removed once the javax.sound bug is
fixed.  We get the delay from the property:

com.sun.speech.freetts.audio.AudioPlayer.closeDelay