Support for TokenGame utterance tests.
Added TokenGame.flite.v1.1.rel which is the output of Flite v1.1 when run against
TokenGame.txt.

Changed the word "lives" in TokenGame.txt to "resides".  Recall that the Flite and FreeTTS
lexicons do not always return the same pronunciation when a word has multiple
pronunciations. Flite and FreeTTS were returning different pronunciations for 'lives'
causing tests to fail. Changing 'lives' to 'resides' alleviates this problem.

Note that the TokenGame utterance test is now run automatically as part of the "make
alltests". It can be run individually by typing:

./utteranceTest.sh ../data/TokenGame.txt ../data/TokenGame.flite.v1.1.rel

from the FreeTTS/tests directory.

Current there are a number of failures when running the test as follows: