Allow 'line.open' to directly call JavaStreamLineListener.update
If the update method is inoked directly from the line.open function,
then the call to openLock.notifyAll() will happen before the call to
openLock.wait() and the code will never recognize that the audio
system is ready for action.

Change openLock into a Semaphore which is released in the update
method and acquired in the openLine method. With this change, it
doesn't matter where the update method is invoked; the Semaphore will
remain signalled until the openLine method invokes the acquire
method.

Signed-off-by: Keith Packard <keithp@keithp.com>